requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
PyMySQL==1.1.1
